+++ owlcms/src/main/java/app/owlcms/data/athlete/Athlete.java
@@ -868,7 +868,7 @@ public class Athlete {
 	 */
 	public void failedLift() {
 		try {
-			getLogger().info("{}no lift for {}", OwlcmsSession.getFopLoggingName(), this.getShortName());
+			getLogger().info("{}no lift for {}", getFopLoggingName(), this.getShortName());
 			final String weight = Integer.toString(-getNextAttemptRequestedWeight());
 			doLift(weight);
 		} catch (Exception e) {
@@ -2100,6 +2100,10 @@ public class Athlete {
 			return OwlcmsSession.getFop();
 		}
 		return this.fop;
+	}
+
+	private String getFopLoggingName() {
+		return FieldOfPlay.getLoggingName(this.fop);
 	}
 
 	@Transient
@@ -4047,7 +4051,7 @@ public class Athlete {
 			validateCleanJerk1ActualLift(cleanJerk1ActualLift);
 		}
 		this.cleanJerk1ActualLift = cleanJerk1ActualLift;
-		getLogger().info("{}{} cleanJerk1ActualLift={}", OwlcmsSession.getFopLoggingName(), this.getShortName(),
+		getLogger().info("{}{} cleanJerk1ActualLift={}", getFopLoggingName(), this.getShortName(),
 				cleanJerk1ActualLift);
 		if (nullIfInvalid(cleanJerk1ActualLift) == null) {
 			this.setCleanJerk1LiftTime(null);
@@ -4072,7 +4076,7 @@ public class Athlete {
 	public void setCleanJerk1Change1(String cleanJerk1Change1) {
 		if ("0".equals(cleanJerk1Change1)) {
 			this.cleanJerk1Change1 = cleanJerk1Change1;
-			getLogger().info("{}{} cleanJerk1Change1={}", OwlcmsSession.getFopLoggingName(), this.getShortName(),
+			getLogger().info("{}{} cleanJerk1Change1={}", getFopLoggingName(), this.getShortName(),
 					cleanJerk1Change1);
 			setCleanJerk1ActualLift("0");
 			return;
@@ -4083,7 +4087,7 @@ public class Athlete {
 		this.cleanJerk1Change1 = cleanJerk1Change1;
 		// validateStartingTotalsRule();
 
-		getLogger().info("{}{} cleanJerk1Change1={}", OwlcmsSession.getFopLoggingName(), this.getShortName(),
+		getLogger().info("{}{} cleanJerk1Change1={}", getFopLoggingName(), this.getShortName(),
 				cleanJerk1Change1);
 	}
 
@@ -4095,7 +4099,7 @@ public class Athlete {
 	public void setCleanJerk1Change2(String cleanJerk1Change2) {
 		if ("0".equals(cleanJerk1Change2)) {
 			this.cleanJerk1Change2 = cleanJerk1Change2;
-			getLogger().info("{}{} cleanJerk1Change2={}", OwlcmsSession.getFopLoggingName(), this.getShortName(),
+			getLogger().info("{}{} cleanJerk1Change2={}", getFopLoggingName(), this.getShortName(),
 					cleanJerk1Change2);
 			setCleanJerk1ActualLift("0");
 			return;
@@ -4106,7 +4110,7 @@ public class Athlete {
 		this.cleanJerk1Change2 = cleanJerk1Change2;
 		// validateStartingTotalsRule();
 
-		getLogger().info("{}{} cleanJerk1Change2={}", OwlcmsSession.getFopLoggingName(), this.getShortName(),
+		getLogger().info("{}{} cleanJerk1Change2={}", getFopLoggingName(), this.getShortName(),
 				cleanJerk1Change2);
 	}
 
@@ -4119,7 +4123,7 @@ public class Athlete {
 		// if ("0".equals(cleanJerk1Declaration)) {
 		// this.cleanJerk1Declaration = cleanJerk1Declaration;
 		// getLogger().info("{}{} cleanJerk1Declaration={}",
-		// OwlcmsSession.getFopLoggingName(), this.getShortName(),
+		// getFopLoggingName(), this.getShortName(),
 		// cleanJerk1Declaration);
 		// setCleanJerk1ActualLift("0");
 		// return;
@@ -4132,7 +4136,7 @@ public class Athlete {
 		// if (zeroIfInvalid(getSnatch1Declaration()) > 0)
 		// // validateStartingTotalsRule();
 
-		getLogger().info("{}{} cleanJerk1Declaration={}", OwlcmsSession.getFopLoggingName(), this.getShortName(),
+		getLogger().info("{}{} cleanJerk1Declaration={}", getFopLoggingName(), this.getShortName(),
 				cleanJerk1Declaration);
 	}
 
@@ -4150,7 +4154,7 @@ public class Athlete {
 			validateCleanJerk2ActualLift(cleanJerk2ActualLift);
 		}
 		this.cleanJerk2ActualLift = cleanJerk2ActualLift;
-		getLogger().info("{}{} cleanJerk2ActualLift={}", OwlcmsSession.getFopLoggingName(), this.getShortName(),
+		getLogger().info("{}{} cleanJerk2ActualLift={}", getFopLoggingName(), this.getShortName(),
 				cleanJerk2ActualLift);
 
 		if (nullIfInvalid(cleanJerk2ActualLift) == null) {
@@ -4176,7 +4180,7 @@ public class Athlete {
 	public void setCleanJerk2Change1(String cleanJerk2Change1) {
 		if ("0".equals(cleanJerk2Change1)) {
 			this.cleanJerk2Change1 = cleanJerk2Change1;
-			getLogger().info("{}{} cleanJerk2Change1={}", OwlcmsSession.getFopLoggingName(), this.getShortName(),
+			getLogger().info("{}{} cleanJerk2Change1={}", getFopLoggingName(), this.getShortName(),
 					cleanJerk2Change1);
 			setCleanJerk2ActualLift("0");
 			return;
@@ -4185,7 +4189,7 @@ public class Athlete {
 			validateCleanJerk2Change1(cleanJerk2Change1);
 		}
 		this.cleanJerk2Change1 = cleanJerk2Change1;
-		getLogger().info("{}{} cleanJerk2Change1={}", OwlcmsSession.getFopLoggingName(), this.getShortName(),
+		getLogger().info("{}{} cleanJerk2Change1={}", getFopLoggingName(), this.getShortName(),
 				cleanJerk2Change1);
 	}
 
@@ -4197,7 +4201,7 @@ public class Athlete {
 	public void setCleanJerk2Change2(String cleanJerk2Change2) {
 		if ("0".equals(cleanJerk2Change2)) {
 			this.cleanJerk2Change2 = cleanJerk2Change2;
-			getLogger().info("{}{} cleanJerk2Change2={}", OwlcmsSession.getFopLoggingName(), this.getShortName(),
+			getLogger().info("{}{} cleanJerk2Change2={}", getFopLoggingName(), this.getShortName(),
 					cleanJerk2Change2);
 			setCleanJerk2ActualLift("0");
 			return;
@@ -4206,7 +4210,7 @@ public class Athlete {
 			validateCleanJerk2Change2(cleanJerk2Change2);
 		}
 		this.cleanJerk2Change2 = cleanJerk2Change2;
-		getLogger().info("{}{} cleanJerk2Change2={}", OwlcmsSession.getFopLoggingName(), this.getShortName(),
+		getLogger().info("{}{} cleanJerk2Change2={}", getFopLoggingName(), this.getShortName(),
 				cleanJerk2Change2);
 	}
 
@@ -4218,7 +4222,7 @@ public class Athlete {
 	public void setCleanJerk2Declaration(String cleanJerk2Declaration) {
 		if ("0".equals(cleanJerk2Declaration)) {
 			this.cleanJerk2Declaration = cleanJerk2Declaration;
-			getLogger().info("{}{} cleanJerk2Declaration={}", OwlcmsSession.getFopLoggingName(), this.getShortName(),
+			getLogger().info("{}{} cleanJerk2Declaration={}", getFopLoggingName(), this.getShortName(),
 					cleanJerk2Declaration);
 			setCleanJerk2ActualLift("0");
 			return;
@@ -4227,7 +4231,7 @@ public class Athlete {
 			validateCleanJerk2Declaration(cleanJerk2Declaration);
 		}
 		this.cleanJerk2Declaration = cleanJerk2Declaration;
-		getLogger().info("{}{} cleanJerk2Declaration={}", OwlcmsSession.getFopLoggingName(), this.getShortName(),
+		getLogger().info("{}{} cleanJerk2Declaration={}", getFopLoggingName(), this.getShortName(),
 				cleanJerk2Declaration);
 	}
 
@@ -4255,7 +4259,7 @@ public class Athlete {
 		}
 
 		this.cleanJerk3ActualLift = cleanJerk3ActualLift;
-		getLogger().info("{}{} cleanJerk3ActualLift={}", OwlcmsSession.getFopLoggingName(), this.getShortName(),
+		getLogger().info("{}{} cleanJerk3ActualLift={}", getFopLoggingName(), this.getShortName(),
 				cleanJerk3ActualLift);
 	}
 
@@ -4282,7 +4286,7 @@ public class Athlete {
 	public void setCleanJerk3Change1(String cleanJerk3Change1) {
 		if ("0".equals(cleanJerk3Change1)) {
 			this.cleanJerk3Change1 = cleanJerk3Change1;
-			getLogger().info("{}{} cleanJerk3Change1={}", OwlcmsSession.getFopLoggingName(), this.getShortName(),
+			getLogger().info("{}{} cleanJerk3Change1={}", getFopLoggingName(), this.getShortName(),
 					cleanJerk3Change1);
 			setCleanJerk3ActualLift("0");
 			return;
@@ -4291,7 +4295,7 @@ public class Athlete {
 			validateCleanJerk3Change1(cleanJerk3Change1);
 		}
 		this.cleanJerk3Change1 = cleanJerk3Change1;
-		getLogger().info("{}{} cleanJerk3Change1={}", OwlcmsSession.getFopLoggingName(), this.getShortName(),
+		getLogger().info("{}{} cleanJerk3Change1={}", getFopLoggingName(), this.getShortName(),
 				cleanJerk3Change1);
 	}
 
@@ -4303,7 +4307,7 @@ public class Athlete {
 	public void setCleanJerk3Change2(String cleanJerk3Change2) {
 		if ("0".equals(cleanJerk3Change2)) {
 			this.cleanJerk3Change2 = cleanJerk3Change2;
-			getLogger().info("{}{} cleanJerk3Change2={}", OwlcmsSession.getFopLoggingName(), this.getShortName(),
+			getLogger().info("{}{} cleanJerk3Change2={}", getFopLoggingName(), this.getShortName(),
 					cleanJerk3Change2);
 			setCleanJerk3ActualLift("0");
 			return;
@@ -4313,7 +4317,7 @@ public class Athlete {
 			validateCleanJerk3Change2(cleanJerk3Change2);
 		}
 		this.cleanJerk3Change2 = cleanJerk3Change2;
-		getLogger().info("{}{} cleanJerk3Change2={}", OwlcmsSession.getFopLoggingName(), this.getShortName(),
+		getLogger().info("{}{} cleanJerk3Change2={}", getFopLoggingName(), this.getShortName(),
 				cleanJerk3Change2);
 	}
 
@@ -4325,7 +4329,7 @@ public class Athlete {
 	public void setCleanJerk3Declaration(String cleanJerk3Declaration) {
 		if ("0".equals(cleanJerk3Declaration)) {
 			this.cleanJerk3Declaration = cleanJerk3Declaration;
-			getLogger().info("{}{} cleanJerk3Declaration={}", OwlcmsSession.getFopLoggingName(), this.getShortName(),
+			getLogger().info("{}{} cleanJerk3Declaration={}", getFopLoggingName(), this.getShortName(),
 					cleanJerk3Declaration);
 			setCleanJerk3ActualLift("0");
 			return;
@@ -4334,7 +4338,7 @@ public class Athlete {
 			validateCleanJerk3Declaration(cleanJerk3Declaration);
 		}
 		this.cleanJerk3Declaration = cleanJerk3Declaration;
-		getLogger().info("{}{} cleanJerk3Declaration={}", OwlcmsSession.getFopLoggingName(), this.getShortName(),
+		getLogger().info("{}{} cleanJerk3Declaration={}", getFopLoggingName(), this.getShortName(),
 				cleanJerk3Declaration);
 	}
 
@@ -4751,7 +4755,7 @@ public class Athlete {
 			validateSnatch1ActualLift(snatch1ActualLift);
 		}
 		this.snatch1ActualLift = snatch1ActualLift;
-		getLogger().info("{}{} snatch1ActualLift={}", OwlcmsSession.getFopLoggingName(), this.getShortName(),
+		getLogger().info("{}{} snatch1ActualLift={}", getFopLoggingName(), this.getShortName(),
 				snatch1ActualLift);
 
 		if (nullIfInvalid(snatch1ActualLift) == null) {
@@ -4777,12 +4781,12 @@ public class Athlete {
 	public void setSnatch1Change1(String snatch1Change1) {
 		if ("0".equals(snatch1Change1)) {
 			this.snatch1Change1 = snatch1Change1;
-			getLogger().info("{}{} snatch1Change1={}", OwlcmsSession.getFopLoggingName(), this.getShortName(),
+			getLogger().info("{}{} snatch1Change1={}", getFopLoggingName(), this.getShortName(),
 					snatch1Change1);
 			setSnatch1ActualLift("0");
 			return;
 		}
-		getLogger().info("{}{} snatch1Change1={}", OwlcmsSession.getFopLoggingName(), this.getShortName(),
+		getLogger().info("{}{} snatch1Change1={}", getFopLoggingName(), this.getShortName(),
 				snatch1Change1);
 		if (isValidation()) {
 			validateSnatch1Change1(snatch1Change1);
@@ -4799,12 +4803,12 @@ public class Athlete {
 	public void setSnatch1Change2(String snatch1Change2) {
 		if ("0".equals(snatch1Change2)) {
 			this.snatch1Change2 = snatch1Change2;
-			getLogger().info("{}{} snatch1Change2={}", OwlcmsSession.getFopLoggingName(), this.getShortName(),
+			getLogger().info("{}{} snatch1Change2={}", getFopLoggingName(), this.getShortName(),
 					snatch1Change2);
 			setSnatch1ActualLift("0");
 			return;
 		}
-		getLogger().info("{}{} snatch1Change2={}", OwlcmsSession.getFopLoggingName(), this.getShortName(),
+		getLogger().info("{}{} snatch1Change2={}", getFopLoggingName(), this.getShortName(),
 				snatch1Change2);
 		if (isValidation()) {
 			try {
@@ -4827,7 +4831,7 @@ public class Athlete {
 			validateSnatch1Declaration(snatch1Declaration);
 		}
 		this.snatch1Declaration = snatch1Declaration;
-		getLogger().info("{}{} snatch1Declaration={}", OwlcmsSession.getFopLoggingName(), this.getShortName(),
+		getLogger().info("{}{} snatch1Declaration={}", getFopLoggingName(), this.getShortName(),
 				snatch1Declaration);
 	}
 
@@ -4845,7 +4849,7 @@ public class Athlete {
 			validateSnatch2ActualLift(snatch2ActualLift);
 		}
 		this.snatch2ActualLift = snatch2ActualLift;
-		getLogger().info("{}{} snatch2ActualLift={}", OwlcmsSession.getFopLoggingName(), this.getShortName(),
+		getLogger().info("{}{} snatch2ActualLift={}", getFopLoggingName(), this.getShortName(),
 				snatch2ActualLift);
 		if (nullIfInvalid(snatch2ActualLift) == null) {
 			this.setSnatch2LiftTime(null);
@@ -4870,7 +4874,7 @@ public class Athlete {
 	public void setSnatch2Change1(String snatch2Change1) {
 		if ("0".equals(snatch2Change1)) {
 			this.snatch2Change1 = snatch2Change1;
-			getLogger().info("{}{} snatch2Change1={}", OwlcmsSession.getFopLoggingName(), this.getShortName(),
+			getLogger().info("{}{} snatch2Change1={}", getFopLoggingName(), this.getShortName(),
 					snatch2Change1);
 			setSnatch2ActualLift("0");
 			return;
@@ -4879,7 +4883,7 @@ public class Athlete {
 			validateSnatch2Change1(snatch2Change1);
 		}
 		this.snatch2Change1 = snatch2Change1;
-		getLogger().info("{}{} snatch2Change1={}", OwlcmsSession.getFopLoggingName(), this.getShortName(),
+		getLogger().info("{}{} snatch2Change1={}", getFopLoggingName(), this.getShortName(),
 				snatch2Change1);
 	}
 
@@ -4891,7 +4895,7 @@ public class Athlete {
 	public void setSnatch2Change2(String snatch2Change2) {
 		if ("0".equals(snatch2Change2)) {
 			this.snatch2Change2 = snatch2Change2;
-			getLogger().info("{}{} snatch2Change2={}", OwlcmsSession.getFopLoggingName(), this.getShortName(),
+			getLogger().info("{}{} snatch2Change2={}", getFopLoggingName(), this.getShortName(),
 					snatch2Change2);
 			setSnatch2ActualLift("0");
 			return;
@@ -4900,7 +4904,7 @@ public class Athlete {
 			validateSnatch2Change2(snatch2Change2);
 		}
 		this.snatch2Change2 = snatch2Change2;
-		getLogger().info("{}{} snatch2Change2={}", OwlcmsSession.getFopLoggingName(), this.getShortName(),
+		getLogger().info("{}{} snatch2Change2={}", getFopLoggingName(), this.getShortName(),
 				snatch2Change2);
 	}
 
@@ -4912,7 +4916,7 @@ public class Athlete {
 	public void setSnatch2Declaration(String snatch2Declaration) {
 		if ("0".equals(snatch2Declaration)) {
 			this.snatch2Declaration = snatch2Declaration;
-			getLogger().info("{}{} snatch2Declaration={}", OwlcmsSession.getFopLoggingName(), this.getShortName(),
+			getLogger().info("{}{} snatch2Declaration={}", getFopLoggingName(), this.getShortName(),
 					snatch2Declaration);
 			setSnatch2ActualLift("0");
 			return;
@@ -4921,7 +4925,7 @@ public class Athlete {
 			validateSnatch2Declaration(snatch2Declaration);
 		}
 		this.snatch2Declaration = snatch2Declaration;
-		getLogger().info("{}{} snatch2Declaration={}", OwlcmsSession.getFopLoggingName(), this.getShortName(),
+		getLogger().info("{}{} snatch2Declaration={}", getFopLoggingName(), this.getShortName(),
 				snatch2Declaration);
 	}
 
@@ -4950,7 +4954,7 @@ public class Athlete {
 		}
 
 		this.snatch3ActualLift = snatch3ActualLift;
-		getLogger().info("{}{} snatch3ActualLift={}", OwlcmsSession.getFopLoggingName(), this.getShortName(),
+		getLogger().info("{}{} snatch3ActualLift={}", getFopLoggingName(), this.getShortName(),
 				snatch3ActualLift);
 	}
 
@@ -4970,7 +4974,7 @@ public class Athlete {
 	public void setSnatch3Change1(String snatch3Change1) {
 		if ("0".equals(snatch3Change1)) {
 			this.snatch3Change1 = snatch3Change1;
-			getLogger().info("{}{} snatch3Change1={}", OwlcmsSession.getFopLoggingName(), this.getShortName(),
+			getLogger().info("{}{} snatch3Change1={}", getFopLoggingName(), this.getShortName(),
 					snatch3Change1);
 			setSnatch3ActualLift("0");
 			return;
@@ -4979,7 +4983,7 @@ public class Athlete {
 			validateSnatch3Change1(snatch3Change1);
 		}
 		this.snatch3Change1 = snatch3Change1;
-		getLogger().info("{}{} snatch3Change1={}", OwlcmsSession.getFopLoggingName(), this.getShortName(),
+		getLogger().info("{}{} snatch3Change1={}", getFopLoggingName(), this.getShortName(),
 				snatch3Change1);
 	}
 
@@ -4991,7 +4995,7 @@ public class Athlete {
 	public void setSnatch3Change2(String snatch3Change2) {
 		if ("0".equals(snatch3Change2)) {
 			this.snatch3Change2 = snatch3Change2;
-			getLogger().info("{}{} snatch3Change2={}", OwlcmsSession.getFopLoggingName(), this.getShortName(),
+			getLogger().info("{}{} snatch3Change2={}", getFopLoggingName(), this.getShortName(),
 					snatch3Change2);
 			setSnatch3ActualLift("0");
 			return;
@@ -5000,7 +5004,7 @@ public class Athlete {
 			validateSnatch3Change2(snatch3Change2);
 		}
 		this.snatch3Change2 = snatch3Change2;
-		getLogger().info("{}{} snatch3Change2={}", OwlcmsSession.getFopLoggingName(), this.getShortName(),
+		getLogger().info("{}{} snatch3Change2={}", getFopLoggingName(), this.getShortName(),
 				snatch3Change2);
 	}
 
@@ -5012,7 +5016,7 @@ public class Athlete {
 	public void setSnatch3Declaration(String snatch3Declaration) {
 		if ("0".equals(snatch3Declaration)) {
 			this.snatch3Declaration = snatch3Declaration;
-			getLogger().info("{}{} snatch3Declaration={}", OwlcmsSession.getFopLoggingName(), this.getShortName(),
+			getLogger().info("{}{} snatch3Declaration={}", getFopLoggingName(), this.getShortName(),
 					snatch3Declaration);
 			setSnatch3ActualLift("0");
 			return;
@@ -5021,7 +5025,7 @@ public class Athlete {
 			validateSnatch3Declaration(snatch3Declaration);
 		}
 		this.snatch3Declaration = snatch3Declaration;
-		getLogger().info("{}{} snatch3Declaration={}", OwlcmsSession.getFopLoggingName(), this.getShortName(),
+		getLogger().info("{}{} snatch3Declaration={}", getFopLoggingName(), this.getShortName(),
 				snatch3Declaration);
 	}
 
@@ -5218,7 +5222,7 @@ public class Athlete {
 	 */
 	public void successfulLift() {
 		try {
-			getLogger().info("{}good lift for {}", OwlcmsSession.getFopLoggingName(), this.getShortName());
+			getLogger().info("{}good lift for {}", getFopLoggingName(), this.getShortName());
 			final String weight = Integer.toString(getNextAttemptRequestedWeight());
 			doLift(weight);
 		} catch (Exception e) {
@@ -5652,13 +5656,13 @@ public class Athlete {
 				referenceWeight, curLift, currentLiftNo_1);
 		// Careful: do not mix one-based numbers with zero-based numbers.
 		if (referenceAttemptNo_1 <= 3 && currentLiftNo_1 == 4) {
-			getLogger().info("{}start of CJ {}", OwlcmsSession.getFopLoggingName(), curLift);
+			getLogger().info("{}start of CJ {}", getFopLoggingName(), curLift);
 			// first attempt for C&J, no check
 			return;
 		}
 
 		if (requestedWeight > referenceWeight) {
-			getLogger().trace("{}{} attempt {}: requested {} > previous {}", OwlcmsSession.getFopLoggingName(), this,
+			getLogger().trace("{}{} attempt {}: requested {} > previous {}", getFopLoggingName(), this,
 					currentLiftNo_1,
 					requestedWeight,
 					referenceWeight);
@@ -5667,7 +5671,7 @@ public class Athlete {
 		}
 
 		if (requestedWeight < referenceWeight) {
-			getLogger().trace("{}requestedWeight {} < referenceWeight {}", OwlcmsSession.getFopLoggingName(),
+			getLogger().trace("{}requestedWeight {} < referenceWeight {}", getFopLoggingName(),
 					requestedWeight,
 					referenceWeight);
 			// someone has already lifted heavier previously
@@ -5722,7 +5726,7 @@ public class Athlete {
 		// logger.debug(" checkChangeVsTimer {} {}", curLift, LoggerUtils.whereFrom());
 
 		Object wi = OwlcmsSession.getAttribute("weighIn");
-		String fopLoggingName = OwlcmsSession.getFopLoggingName();
+		String fopLoggingName = getFopLoggingName();
 		if (wi == this) {
 			// current athlete being weighed in
 			getLogger().debug("{}weighin {}", fopLoggingName, wi);
@@ -5754,7 +5758,7 @@ public class Athlete {
 		if (declaration == null || declaration.isBlank()) {
 			// there was no declaration made in time
 			this.logger./**/warn("{}{} change without declaration (not owning clock)",
-					OwlcmsSession.getFopLoggingName(),
+					getFopLoggingName(),
 					this.getShortName());
 			throw new RuleViolationException.MustDeclareFirst(this, clock);
 		}
@@ -5765,7 +5769,7 @@ public class Athlete {
 	private void checkSameProgression(LiftOrderInfo reference, Integer requestedWeight, int currentProgression,
 			int referenceProgression) {
 		long start = System.currentTimeMillis();
-		String fopLoggingName = OwlcmsSession.getFopLoggingName();
+		String fopLoggingName = getFopLoggingName();
 		getLogger().trace("{}currentProgression {} == referenceProgression {}", fopLoggingName, currentProgression,
 				referenceProgression);
 		if (this.getStartNumber() > 0) {
@@ -5799,7 +5803,7 @@ public class Athlete {
 	private void checkSameWeightAsReference(LiftOrderInfo reference, Integer requestedWeight, int referenceWeight,
 			int referenceAttemptNo, int currentLiftNo) {
 		long start = System.currentTimeMillis();
-		String fopLoggingName = OwlcmsSession.getFopLoggingName();
+		String fopLoggingName = getFopLoggingName();
 		getLogger().trace("{}requestedWeight {} == referenceWeight {}",
 				fopLoggingName, requestedWeight, referenceWeight);
 		// asking for same weight as previous lifter, cannot be a lower attempt number
@@ -5886,10 +5890,10 @@ public class Athlete {
 			int clock, int initialTime) {
 		if ((declaration != null && !declaration.isBlank()) && (change1 == null || change1.isBlank())
 				&& (change2 == null || change2.isBlank())) {
-			this.logger.trace("{}{} declaration accepted (not owning clock)", OwlcmsSession.getFopLoggingName(),
+			this.logger.trace("{}{} declaration accepted (not owning clock)", getFopLoggingName(),
 					this.getShortName());
 		} else {
-			this.logger.trace("{}{} change accepted (not owning clock)", OwlcmsSession.getFopLoggingName(),
+			this.logger.trace("{}{} change accepted (not owning clock)", getFopLoggingName(),
 					this.getShortName());
 		}
 	}
@@ -5904,21 +5908,21 @@ public class Athlete {
 		if ((change1 == null || change1.isBlank()) && (change2 == null || change2.isBlank())) {
 			// validate declaration
 			if (clock < initialTime - Competition.athleteTimerFinalWarning) {
-				this.logger./**/warn("{}{} late declaration denied ({})", OwlcmsSession.getFopLoggingName(),
+				this.logger./**/warn("{}{} late declaration denied ({})", getFopLoggingName(),
 						this.getShortName(),
 						clock / 1000.0);
 				throw new RuleViolationException.LateDeclaration(this, clock);
 			}
-			this.logger.debug("{}{}valid declaration", OwlcmsSession.getFopLoggingName(), this.getShortName(),
+			this.logger.debug("{}{}valid declaration", getFopLoggingName(), this.getShortName(),
 					clock / 1000.0);
 		} else {
 			if (clock < Competition.athleteTimerFinalWarning) {
 				this.logger./**/warn("{}{} late change denied after final warning ({})",
-						OwlcmsSession.getFopLoggingName(),
+						getFopLoggingName(),
 						this.getShortName(), clock / 1000.0);
 				throw new RuleViolationException.MustChangeBeforeFinalWarning(this, clock);
 			}
-			this.logger.debug("{}change before final warning", OwlcmsSession.getFopLoggingName(), clock);
+			this.logger.debug("{}change before final warning", getFopLoggingName(), clock);
 		}
 	}
 
@@ -5948,7 +5952,7 @@ public class Athlete {
 		}
 
 		Object wi = OwlcmsSession.getAttribute("weighIn");
-		String fopLoggingName = OwlcmsSession.getFopLoggingName();
+		String fopLoggingName = getFopLoggingName();
 		if (wi == this) {
 			// athlete being weighed in
 			getLogger().trace("{}weighin {}", fopLoggingName, wi);
@@ -6170,7 +6174,7 @@ public class Athlete {
 		loi.setCumulativeProgression(this.getCumulativeProgression(nextAttemptRequestedWeight));
 		loi.setStartNumber(this.getStartNumber());
 		loi.setLotNumber(this.getLotNumber());
-		// getLogger().trace("{}clockOwner: {}", OwlcmsSession.getFopLoggingName(),
+		// getLogger().trace("{}clockOwner: {}", getFopLoggingName(),
 		// loi);
 		return loi;
 	}
@@ -6484,7 +6488,7 @@ public class Athlete {
 		}
 		long start = System.currentTimeMillis();
 		// getLogger().debug("{}{} validateDeclaration {} {} {} from {}",
-		// OwlcmsSession.getFopLoggingName(), this,
+		// getFopLoggingName(), this,
 		// declaration, change1, change2, LoggerUtils.stackTrace());
 		int newVal = zeroIfInvalid(declaration);
 		int iAutomaticProgression = zeroIfInvalid(automaticProgression);